# Conflicts:
#	src/main/kotlin/senior/joinu/candid/serialize/TypeSers.kt
#	src/main/kotlin/senior/joinu/leb128/Leb128.kt

+++ src/main/kotlin/senior/joinu/candid/serialize/ValueSers.kt
@@ -260,18 +260,18 @@ object TextValueSer : ValueSer<String> {
     override fun calcSizeBytes(value: String): Int {
         val strBytes = value.toByteArray(StandardCharsets.UTF_8)
 
-        return Leb128.sizeSigned(strBytes.size) + strBytes.size
+        return Leb128.sizeSigned(strBytes.size.toLong()) + strBytes.size
     }
 
     override fun ser(buf: ByteBuffer, value: String) {
         val strBytes = value.toByteArray(StandardCharsets.UTF_8)
-        Leb128.writeUnsigned(buf, strBytes.size)
+        Leb128.writeUnsigned(buf, strBytes.size.toLong())
         buf.put(strBytes)
     }
 
     override fun deser(buf: ByteBuffer): String {
         val strLength = Leb128.readUnsigned(buf)
-        val strBytes = ByteArray(strLength)
+        val strBytes = ByteArray(strLength.toInt())
         buf.get(strBytes)
 
         return strBytes.toString(StandardCharsets.UTF_8)
@@ -362,11 +362,11 @@ class OptValueSer<T : Any>(val innerSer: ValueSer<T>) : ValueSer<T?> {
 
 class VecValueSer<T>(val innerSer: ValueSer<T>) : ValueSer<List<T>> {
     override fun calcSizeBytes(value: List<T>): Int {
-        return Leb128.sizeUnsigned(value.size) + value.map { innerSer.calcSizeBytes(it) }.sum()
+        return Leb128.sizeUnsigned(value.size.toLong()) + value.map { innerSer.calcSizeBytes(it) }.sum()
     }
 
     override fun ser(buf: ByteBuffer, value: List<T>) {
-        Leb128.writeUnsigned(buf, value.size)
+        Leb128.writeUnsigned(buf, value.size.toLong())
 
         for (entry in value) {
             innerSer.ser(buf, entry)
@@ -385,18 +385,18 @@ class VecValueSer<T>(val innerSer: ValueSer<T>) : ValueSer<List<T>> {
 
 object BlobValueSer : ValueSer<ByteArray> {
     override fun calcSizeBytes(value: ByteArray): Int {
-        return Leb128.sizeUnsigned(value.size) + value.size
+        return Leb128.sizeUnsigned(value.size.toLong()) + value.size
     }
 
     override fun ser(buf: ByteBuffer, value: ByteArray) {
-        Leb128.writeUnsigned(buf, value.size)
+        Leb128.writeUnsigned(buf, value.size.toLong())
 
         buf.put(value)
     }
 
     override fun deser(buf: ByteBuffer): ByteArray {
         val size = Leb128.readUnsigned(buf)
-        val bytes = ByteArray(size)
+        val bytes = ByteArray(size.toInt())
         buf.get(bytes)
 
         return bytes
